+++ lib/clean_architecture/serializers/json_response_from_result.rb
@@ -2,6 +2,8 @@
 
 require 'dry-monads'
 require 'dry/matcher/result_matcher'
+require 'clean_architecture/entities/failure_details'
+require 'clean_architecture/queries/http_failure_code'
 require 'clean_architecture/queries/http_success_code'
 
 module CleanArchitecture
@@ -25,8 +27,10 @@ module CleanArchitecture
       def http_status_code
         Dry::Matcher::ResultMatcher.call(@result) do |matcher|
           matcher.success { Queries::HttpSuccessCode.new(@http_method).to_sym }
-          matcher.failure do
-            if result_is_authorization_failure?
+          matcher.failure do |failure_value|
+            if failure_value.is_a?(Entities::FailureDetails)
+              Queries::HttpFailureCode.new(failure_value.type).to_sym
+            elsif result_is_authorization_failure?
               :unauthorized
             else
               :expectation_failed
@@ -38,7 +42,7 @@ module CleanArchitecture
       def json
         Dry::Matcher::ResultMatcher.call(@result) do |matcher|
           matcher.success { success_payload }
-          matcher.failure { |failure_message| error_payload(failure_message) }
+          matcher.failure { |failure_value| error_payload(failure_value) }
         end
       end
 
@@ -49,10 +53,11 @@ module CleanArchitecture
         }
       end
 
-      def error_payload(failure_message)
+      def error_payload(failure_value)
+        msg = failure_value.is_a?(Entities::FailureDetails) ? failure_value.message : failure_value
         {
           jsonapi: json_api_version_hash,
-          errors: [failure_message]
+          errors: [msg]
         }
       end
 
@@ -61,7 +66,8 @@ module CleanArchitecture
       end
 
       def result_is_authorization_failure?
-        !@result.failure.index('Unauthorized: ').nil?
+        msg = @result.failure
+        msg.is_a?(String) && !msg.index('Unauthorized: ').nil?
       end
     end
   end