+++ spec/lib/clean_architecture/serializers/json_response_from_result_spec.rb
@@ -61,6 +61,24 @@ module CleanArchitecture
             )
           end
         end
+
+        context do
+          let(:failure_details) do
+            Entities::FailureDetails.new(
+              message: 'Something bad happened!',
+              other_properties: {},
+              type: 'error'
+            )
+          end
+          let(:result) { Dry::Monads::Failure(failure_details) }
+
+          specify do
+            expect(to_h).to eq(
+              status: :internal_server_error,
+              json: { jsonapi: { version: '1.0' }, errors: ['Something bad happened!'] }
+            )
+          end
+        end
       end
     end
   end