+++ spec/lib/clean_architecture/strategies/actor_gets_authorized_then_does_work_spec.rb
@@ -33,7 +33,7 @@ module CleanArchitecture
 
           specify do
             expect(result).to be_a Dry::Monads::Failure
-            expect(result.failure).to eq 'Unauthorized'
+            expect(result.failure).to eq described_class::UNAUTHORIZED_FAILURE_DETAILS
           end
         end
       end